Backend spring boot apirest modificaciones tipado id y readonli

+++ spring-boot-backend-apirest/src/main/java/com/capgemini/springboot/backend/apirest/models/entity/Cliente.java
@@ -20,7 +20,7 @@ public class Cliente implements Serializable{
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String nombre;
 	//@Column() solo cuando el nombre en la tabla es distito al que se pone aquí
 	private String apellido;
@@ -34,12 +34,11 @@ public class Cliente implements Serializable{
 		createAt= new Date();
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
